split + corrections de la partie 1

+++ fonctions/ft_memcpy.c
@@ -2,8 +2,8 @@
 
 void	*ft_memcpy(void *dest, const void *src, size_t n)
 {
-	int	i;
-	char	*dest_ptr;
+	size_t		i;
+	char		*dest_ptr;
 	const char	*src_ptr;
 
 	i = 0;
